Completed the Widgets, Window, Desktop classes

+++ include/drivers/vga/VGA.hpp
@@ -62,8 +62,8 @@ class VideoGraphicsArray {
 
         virtual bool supportsMode(uint32_t width, uint32_t height, uint32_t depth);
         virtual bool setMode(uint32_t width, uint32_t height, uint32_t depth);
-        virtual void setPixel(uint8_t x, uint8_t y, uint8_t depth);
-        virtual void setPixel(uint32_t x, uint32_t y, uint8_t r, uint8_t g, uint8_t b);
+        virtual void setPixel(int32_t x, int32_t y, uint8_t depth);
+        virtual void setPixel(int32_t x, int32_t y, uint8_t r, uint8_t g, uint8_t b);
         virtual void fillRect(uint32_t x, uint32_t y, uint32_t w, uint32_t h, uint8_t r, uint8_t g, uint8_t b);
 };
